handleSubmitをlib/click.tsに移動
